requirements.txt
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
nltk>=3.8.0
scikit-learn>=1.2.0
pyspark>=3.3.0
textblob>=0.17.0
matplotlib>=3.6.0
yfinance>=0.2.0
python-dotenv>=0.21.0
streamlit>=1.28.0
plotly>=5.14.0
joblib>=1.2.0
lxml>=4.9.0
